Moving to back to resources. Resources or idempotent operations -- what's a better primitive?
If you go with idempotent operations, it's hard to build resources out of them
unless you can extend the checks in an operation. Say for example we have `rm
<path>'. It's run if the <path> is there. Later, we want to have some
fanciness to `rm'; like, only `rm' if there is a symlink link at <path> that
points to `/dev/null'. We approach this as a "filesystem node" resource that
implies a lot of idempotent operations and checks. Now how do we add extra
checks to the `rm' so that it does not simply nuke the path if it is there?